Removed print statements. extracted out function to a utility and imported back into the files.
