+++ utils/UtilityFunctions.js
@@ -1,3 +1,5 @@
+import { months } from "@/config/index";
+
 export function cleanPhone(phoneNumber) {
   const regexPattern = /[^0-9]+/g;
   const newPhoneNumber = phoneNumber.replace(regexPattern, "");
@@ -15,12 +17,27 @@ export function cleanPhone(phoneNumber) {
     desiredPhoneFormat = `${countryCode} (${areaCode}) ${cityCode}-${lastFour}`;
   }
 
-  console.log(desiredPhoneFormat);
-  console.log(newPhoneNumber);
-
   if (desiredPhoneFormat.length > newPhoneNumber.length) {
     return desiredPhoneFormat;
   } else {
     return newPhoneNumber;
   }
+}
+
+export function formatBlogDate(dateTime) {
+  let dateStr = "";
+
+  try {
+    let dtSplit = dateTime.split("-");
+    let dateSplit = dtSplit[2].split("T");
+    let month = months[parseInt(dtSplit[1] - 1)];
+    let day = dateSplit[0];
+    let year = dtSplit[0];
+
+    dateStr = `${month} ${day}, ${year}`;
+  } catch (error) {
+    // Todo: add logging
+  }
+
+  return dateStr;
 }
